requirements.txt
# CORE LIBS
openai>=1.0
pydub
ffmpeg-python
python-dotenv
